Refactoring node functions to be separate
